refactor: STEP-4 피드백 반영
name 검증을 Car 클래스로 이동
자동차의 전진/정지를 결정하는 RaceRule 인터페이스 추가
RacingGame 클래스 일급 컬렉션으로 변경
