+++ src/main/java/step3/GameLauncher.java
@@ -1,15 +1,20 @@
 package step3;
 
+import step3.gameobject.RacingGame;
+import step3.gameobject.rule.DefaultRaceRule;
+import step3.ui.InputView;
+import step3.ui.ResultView;
+
+import java.util.List;
+
 public class GameLauncher {
 
     public static void main(String[] args) {
-        RacingGame racingGame = new RacingGame();
-
-        String namesOfCars = InputView.getNameOfCars();
+        List<String> nameList = InputView.getNameOfCars();
         int count = InputView.getCount();
         ResultView.printResultMessage();
 
-        racingGame.preparingGame(namesOfCars, count);
+        RacingGame racingGame = new RacingGame(nameList, new DefaultRaceRule());
         for (int i = 0; i < count; i++) {
             ResultView.printEachResult(racingGame.race());
         }
