+++ src/main/java/step3/gameobject/rule/RaceRule.java
@@ -1,0 +1,7 @@
+package step3.gameobject.rule;
+
+public interface RaceRule {
+
+    boolean race();
+
+}
