+++ src/main/java/step3/ui/ResultView.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.ui;
 
 import java.util.List;
 
@@ -6,7 +6,7 @@ public class ResultView {
 
     private static final String RESULT_MESSAGE = "실행 결과";
     private static final String DELIMITER_FOR_WINNER = ",";
-    private static final String FINAL_WINNER_MESSAGE = "가 최종 우승했습니다.";
+    private static final String FINAL_WINNER_MESSAGE = " 가 최종 우승했습니다.";
 
     public static void newLine() {
         System.out.println();
