+++ src/test/java/step3/CarTest.java
@@ -3,6 +3,8 @@ package step3;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import step3.gameobject.Car;
+import step3.rule.StubbingTrueRule;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -13,11 +15,11 @@ public class CarTest {
     @ValueSource(ints = {1, 2, 3, 4, 10})
     public void plusTraceTest(int count) {
         String name = "TEST";
-        Car car = new Car(name);
+        Car car = new Car(name, new StubbingTrueRule());
         StringBuilder expected = new StringBuilder("-");
 
         for (int i = 0; i < count; i++) {
-            car.goOrStop(true);
+            car.goOrStop();
             expected.append("-");
         }
