+++ src/test/java/step3/rule/FirstRacerWinRule.java
@@ -1,0 +1,16 @@
+package step3.rule;
+
+import step3.gameobject.rule.RaceRule;
+
+public class FirstRacerWinRule implements RaceRule {
+    boolean isFirstRacer = true;
+
+    @Override
+    public boolean race() {
+        if (isFirstRacer) {
+            isFirstRacer = false;
+            return true;
+        }
+        return false;
+    }
+}
