+++ src/test/java/step3/rule/StubbingTrueRule.java
@@ -1,0 +1,12 @@
+package step3.rule;
+
+import step3.gameobject.rule.RaceRule;
+
+public class StubbingTrueRule implements RaceRule {
+
+    @Override
+    public boolean race() {
+        return true;
+    }
+
+}
